create node type

+++ src/lib.rs
@@ -8,12 +8,13 @@ use crate::solve_matrix::{SolvingMatrix, SquareNumber};
 // TODO: implement methods here so that JS doesn't have to extract?
 #[wasm_bindgen]
 pub struct FinalSudokuBoard {
-    init_grid: Vec<u16>,
-    solved_grid: Vec<u16>,
+    init_grid: Vec<u8>,
+    solved_grid: Vec<u8>,
 }
 
 #[wasm_bindgen]
-pub fn generate_sudoku(input_sudoku_size: u16) -> Option<FinalSudokuBoard> {
+pub fn generate_sudoku(input_sudoku_size: u8) -> Option<FinalSudokuBoard> {
+    // TODO: choose max size allowed
     let valid_sudoku_size: SquareNumber = input_sudoku_size.try_into().ok()?;
     let mut matrix = SolvingMatrix::new(valid_sudoku_size);
     // SOLVE HERE
